fsm: separate enterprise methods into their own _oss file.
